requirements.txt
apache-beam[gcp]
google-cloud-bigquery
pandas~=1.3.4
numpy~=1.20.3
GoogleAppEngineCloudStorageClient~=1.9.22.1